% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genealogy.R
\name{gather_segments}
\alias{gather_segments}
\alias{make_gene_genealogy}
\alias{count_uncoalesced}
\title{Functions to generate gene genealogy from sample family table}
\usage{
gather_segments(samples)

make_gene_genealogy(segments)

count_uncoalesced(genealogy)
}
\arguments{
\item{samples}{A \code{sample_family} data.frame.}

\item{segments}{An output from \code{\link[=gather_segments]{gather_segments()}}.}

\item{genealogy}{An output from \code{\link[=make_gene_genealogy]{make_gene_genealogy()}}.}
}
\description{
\code{\link[=gather_segments]{gather_segments()}} transforms an individual-based \code{sample_family} into
a segment-based table.
}
\details{
\code{\link[=make_gene_genealogy]{make_gene_genealogy()}} generates a random gene genealogy from a
segment-based table.

\code{\link[=count_uncoalesced]{count_uncoalesced()}} counts uncoalesced roots of the samples.
}
