% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsp2.R
\name{as_hsp2}
\alias{as_hsp2}
\alias{write_hsp2}
\alias{read_hsp2}
\title{Count half-sib pairs in HSP2 format}
\usage{
as_hsp2(samples)

write_hsp2(x, path = "hsp2.tsv")

read_hsp2(path)
}
\arguments{
\item{samples}{A \code{sample_family} data.frame of \code{\link[=tekka]{tekka()}} result.}

\item{x}{An outcome of \code{\link[=as_hsp2]{as_hsp2()}}.}

\item{path}{A file name or connection to write to.}
}
\value{
A data.frame of "hsp2" subclass with the following columns:
\itemize{
\item \code{cohort_i}, \code{cohort_j}: birth year of samples
\item \code{capture_age_i}, \code{capture_age_j}: of samples
\item \code{location_i}, \code{location_j}: of sampling
\item \code{comps}: the number of possible comparisons between group \emph{i} and \emph{j}.
All samples are compared pairwise, excluding self-comparisons.
The total number of comparisons equals \\(n \\choose 2\\),
where \emph{n} is the total sample size.
\item \code{hsps}: the count of half-sib pairs observed between group \emph{i} and \emph{j}
}
}
\description{
The extended version of HSP format with additional columns for grouping.
}
\examples{
set.seed(666)
result = tekka("-y25 -l2 --sa 2,2 --sj 2,2")
samples = result$sample_family[[1L]]
as_hsp2(samples)
}
\seealso{
\code{\link[=as_hsp]{as_hsp()}} for the original HSP format.

\code{\link[=find_kinship]{find_kinship()}} to count kinship within samples.
}
