% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinship.R
\name{find_kinship}
\alias{find_kinship}
\title{Analyze kinship with igraph}
\usage{
find_kinship(samples, order = 4L, experimental = FALSE)
}
\arguments{
\item{samples}{sample_family}

\item{order}{integer}

\item{experimental}{boolean}
}
\description{
Analyze kinship with igraph
}
\details{
\code{\link[=find_kinship]{find_kinship()}} finds kinship below given order.
}
