% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbinom.R
\name{plot_nbinom}
\alias{plot_nbinom}
\title{Test negative_binomial_distribution in tekka simulation}
\usage{
plot_nbinom(n = 10000L)
}
\arguments{
\item{n}{number of observations.  If \code{length(n) > 1}, the length
    is taken to be the number required.}
}
\description{
\code{cpp_rnbinom()} generates nbinom random numbers.
}
\details{
\code{plot_nbinom()} draws distributions with various parameter sets.
}
\seealso{
stats::rnbinom
}
