% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{migration_matrices}
\alias{migration_matrices}
\alias{natural_mortality}
\alias{fishing_mortality}
\alias{weight_for_age}
\title{Parameter getter}
\usage{
migration_matrices()

natural_mortality()

fishing_mortality()

weight_for_age()
}
\description{
Parameter getter
}
