% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{plot_parameters_json}
\alias{plot_parameters_json}
\title{Visualize parameter json file}
\usage{
plot_parameters_json(obj = default_parameters_json())
}
\arguments{
\item{obj}{jsonlite object}
}
\description{
Visualize parameter json file
}
\examples{
plot_parameters_json()
}
