% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{as_pop}
\alias{as_pop}
\alias{write_pop}
\title{Convert a result data frame to POP format}
\usage{
as_pop(.tbl, min_adult_age = 5L)

write_pop(x, path = "pop.txt")
}
\arguments{
\item{.tbl}{result data frame}

\item{min_adult_age}{integer}

\item{x}{A data frame to write to disk}

\item{path}{Path or connection to write to.}
}
\value{
tibble
}
\description{
Convert a result data frame to POP format

Write a POP data frame to a file
}
