% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop2.R
\name{as_pop2}
\alias{as_pop2}
\alias{write_pop2}
\alias{read_pop2}
\title{Count parent-offspring pairs in POP2 format}
\usage{
as_pop2(samples)

write_pop2(x, path = "pop2.tsv")

read_pop2(path)
}
\arguments{
\item{samples}{A \code{sample_family} data.frame of \code{\link[=tekka]{tekka()}} result.}

\item{x}{An outcome of \code{\link[=as_pop2]{as_pop2()}}}

\item{path}{A file name or connection to write to.}
}
\value{
A data.frame of "pop2" subclass with the following columns:
\itemize{
\item \code{cohort_parent}, \code{cohort_offspring}: birth year of samples
\item \code{capture_age_parent}, \code{capture_age_offspring}
\item \code{location_parent}, \code{location_offspring}: of sampling
\item \code{pops}: the count of parent-offspring pairs observed within samples
\item \code{comps}: the number of possible comparisons,
including only pairs where one sample is younger than the other.
}
}
\description{
The extended version of POP format with additional columns for grouping.
}
\examples{
set.seed(666)
result = tekka("-y25 -l2 --sa 2,2 --sj 2,2")
samples = result$sample_family[[1L]]
as_pop2(samples)
}
\seealso{
\code{\link[=as_pop]{as_pop()}} for the original POP format.

\code{\link[=find_kinship]{find_kinship()}} to count kinship within samples.
}
