% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_tekka}
\alias{read_tekka}
\title{Read results from \code{tekka}}
\usage{
read_tekka(indir = getwd())
}
\arguments{
\item{indir}{An output directory of \code{tekka}.}
}
\value{
A one-row data.frame with the configuration parameters and results.
There are two result columns:
\itemize{
\item \code{sample_family}: a list of data.frames with samples and their ancestors.
\item \code{demography}: a list of data.frames with population demography.
}
}
\description{
The \code{tekka} simulator writes results and configuration files to a directory.
This function reads them into a one-row data.frame.
}
\seealso{
\code{\link[=tekka]{tekka()}} for example usage.
}
