% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampled.R
\name{count_sampled}
\alias{count_sampled}
\title{Count actually sampled individuals}
\usage{
count_sampled(samples)
}
\arguments{
\item{samples}{A \code{sample_family} data.frame of \code{\link[=tekka]{tekka()}} result.}
}
\value{
A data.frame with the following columns:
\itemize{
\item \code{capture_year}
\item \code{location}
\item \code{adult}
\item \code{juvenile}
}
}
\description{
Count actually sampled individuals
}
\examples{
set.seed(666)
result = tekka("-y25 -l2 --sa 2,2 --sj 2,2")
samples = result$sample_family[[1L]]
count_sampled(samples)
}
