% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{gather_chromosome}
\alias{gather_chromosome}
\alias{sprinkle_mutations}
\alias{accumulate_mutations}
\alias{complete_genotype}
\alias{spread_genotype}
\alias{make_snp}
\title{row chromosome}
\usage{
gather_chromosome(.tbl)

sprinkle_mutations(.tbl, lambda)

accumulate_mutations(.tbl)

complete_genotype(.tbl)

spread_genotype(.tbl)

make_snp(.tbl, lambda, rho = 0)
}
\arguments{
\item{.tbl}{result tibble}

\item{lambda}{mutation rate per haploid per generation}

\item{rho}{recombination rate per generation}
}
\value{
tibble
}
\description{
row chromosome

choose random mutation positions [0, 1)

trace back ancestors to collect mutation positions

complete genotypes for all the positions

spread data frame to wide format

shortcut to make snp table
}
