% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tekkamaki}
\alias{tekkamaki}
\alias{tekkamaki-package}
\title{tekkamaki: Individual-based simulator of pacific bluefin tuna}
\description{
tekkamaki: Individual-based simulator of pacific bluefin tuna
}
