% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/blackthunnus.R
\name{cpp_blackthunnus}
\alias{cpp_blackthunnus}
\alias{blackthunnus}
\title{Run C++ simulation}
\usage{
cpp_blackthunnus(args)

blackthunnus(args = character(0L))
}
\arguments{
\item{args}{command line arguments as a string vector}
}
\value{
tibble
}
\description{
Run C++ simulation

Run C++ simulation
}
