% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genealogy.R
\name{make_gene_genealogy}
\alias{make_gene_genealogy}
\alias{gather_segments}
\alias{count_uncoalesced}
\alias{annotate_sampled}
\title{Generate gene genealogy from samples}
\usage{
make_gene_genealogy(samples)

gather_segments(samples)

count_uncoalesced(genealogy)

annotate_sampled(genealogy)
}
\arguments{
\item{samples}{A data.frame: \code{sample_family} or its tranfromation by \code{\link[=gather_segments]{gather_segments()}}.
Using the latter will improve performance when the function is called many times,
e.g., in \code{\link[=make_snp]{make_snp()}}.}

\item{genealogy}{An output from \code{\link[=make_gene_genealogy]{make_gene_genealogy()}}.}
}
\value{
An igraphlite object with \code{genealogy} subclass.

\code{\link[=gather_segments]{gather_segments()}} transforms an individual-based \code{sample_family}
into a segment-based table.

\code{\link[=count_uncoalesced]{count_uncoalesced()}} returns the count of uncoalesced roots of the genealogy.

\code{\link[=annotate_sampled]{annotate_sampled()}} adds the "sampled" edge attribute to the genealogy:
\code{TRUE} if sampled, \code{FALSE} if upstream of samples, \code{NA} if unrelated.
}
\description{
An individual-based family tree stored in \code{sample_family} is transformed
into a gene-based table with \code{\link[=gather_segments]{gather_segments()}}.
Then, the ancestry of each gene/segment is randomly assigned in \code{\link[=make_gene_genealogy]{make_gene_genealogy()}}.
Sampled segments and their ancestors can be annotated with \code{\link[=annotate_sampled]{annotate_sampled()}}.
}
\examples{
set.seed(666)
result = tekka("-y20 -l2 --sa 2,2 --sj 2,2")
samples = result$sample_family[[1L]]
segments = gather_segments(samples)
segments

genealogy = make_gene_genealogy(segments)
genealogy

count_uncoalesced(genealogy)

annotate_sampled(genealogy)
}
\seealso{
\code{\link[=make_snp]{make_snp()}} for SNP generation using genealogy.
}
