% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genealogy.R
\name{make_gene_genealogy}
\alias{make_gene_genealogy}
\alias{gather_segments}
\alias{count_uncoalesced}
\alias{annotate_sampled}
\title{Functions to generate gene genealogy from sample family table}
\usage{
make_gene_genealogy(samples)

gather_segments(samples)

count_uncoalesced(genealogy)

annotate_sampled(genealogy)
}
\arguments{
\item{samples}{A data.frame: \code{sample_family} or output from \code{\link[=gather_segments]{gather_segments()}}.}

\item{genealogy}{An output from \code{\link[=make_gene_genealogy]{make_gene_genealogy()}}.}
}
\description{
\code{\link[=make_gene_genealogy]{make_gene_genealogy()}} generates a random gene genealogy from a table.
}
\details{
\code{\link[=gather_segments]{gather_segments()}} transforms an individual-based \code{sample_family} into
a segment-based table.

\code{\link[=count_uncoalesced]{count_uncoalesced()}} counts uncoalesced roots of the samples.

\code{\link[=annotate_sampled]{annotate_sampled()}} adds "sampled" edge attribute: \code{TRUE} if sampled,
\code{FALSE} if upstream of samples, \code{NA} if unrelated.
}
