% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{as_igraph.sample_family}
\alias{as_igraph.sample_family}
\title{Convert genealogy to igraph}
\usage{
\method{as_igraph}{sample_family}(x, ...)
}
\arguments{
\item{x}{\code{sample_family} table}

\item{...}{Additional arguments passed to \code{\link[igraphlite:conversion-impl]{igraphlite::graph_from_data_frame()}}.}
}
\description{
Convert genealogy to igraph
}
\details{
\code{as_igraph} converts a result to igraph.
}
