% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{as_igraph.sample_family}
\alias{as_igraph.sample_family}
\title{Convert genealogy to igraph}
\usage{
\method{as_igraph}{sample_family}(x)
}
\arguments{
\item{x}{\code{sample_family} table}
}
\description{
Convert genealogy to igraph
}
\details{
\code{as_igraph} converts a result to igraph.
}
