% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{as_igraph}
\alias{as_igraph}
\alias{find_kinship}
\title{Functions for igraph class}
\usage{
as_igraph(.tbl)

find_kinship(.tbl, order = 4L)
}
\arguments{
\item{.tbl}{sample_family}

\item{order}{integer}
}
\description{
Functions for igraph class
}
\details{
\code{as_igraph} converts a result to igraph.

\code{find_kinship} finds kinship below given order.
}
