% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{as_igraph.sample_family}
\alias{as_igraph.sample_family}
\alias{find_kinship}
\title{Functions for igraph class}
\usage{
\method{as_igraph}{sample_family}(.tbl)

find_kinship(.tbl, order = 4L, experimental = FALSE)
}
\arguments{
\item{.tbl}{sample_family}

\item{order}{integer}

\item{experimental}{boolean}
}
\description{
Functions for igraph class
}
\details{
\code{as_igraph} converts a result to igraph.

\code{find_kinship} finds kinship below given order.
}
