% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsp.R
\name{as_hsp}
\alias{as_hsp}
\alias{write_hsp}
\title{HSP format}
\usage{
as_hsp(.tbl)

write_hsp(x, path = "hsp.txt")
}
\arguments{
\item{.tbl}{A result data frame}

\item{x}{An outcome of as_hsp()}

\item{path}{A file name or connection to write to}
}
\description{
\code{as_hsp} converts a result data frame to HSP format.

\code{write_hsp} writes a HSP data frame to a file.
}
