% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsp.R
\name{as_hsp}
\alias{as_hsp}
\alias{write_hsp}
\title{Convert a result data frame to HSP format}
\usage{
as_hsp(.tbl)

write_hsp(x, path = "hsp.txt")
}
\arguments{
\item{.tbl}{result data frame}

\item{x}{A data frame to write to disk}

\item{path}{Path or connection to write to.}
}
\value{
tibble
}
\description{
Convert a result data frame to HSP format

Write a HSP data frame to a file
}
