% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsp2.R
\name{as_hsp2}
\alias{as_hsp2}
\alias{write_hsp2}
\alias{read_hsp2}
\title{Extended HSP format}
\usage{
as_hsp2(samples)

write_hsp2(x, path = "hsp2.tsv")

read_hsp2(path)
}
\arguments{
\item{samples}{A \code{sample_family} data.frame of \code{\link[=tekka]{tekka()}} result.}

\item{x}{An outcome of \code{\link[=as_hsp2]{as_hsp2()}}.}

\item{path}{A file name or connection to write to.}
}
\value{
A data.frame with "hsp2" class and the following columns:
\itemize{
\item \code{cohort_i}, \code{cohort_j}: birth year of samples
\item \code{capture_age_i}, \code{capture_age_j}: of samples
\item \code{location_i}, \code{location_j}: of sampling
\item \code{comps}: the number of possible comparisons between group \emph{i} and \emph{j}
\item \code{hsps}: the number of half-sib pairs between group \emph{i} and \emph{j}
}
}
\description{
Half-sib pairs are counted between samples grouped by birth year, capture age,
and sampled location.
}
\details{
\code{\link[=as_hsp2]{as_hsp2()}} converts a result data frame to HSP format.

\code{\link[=write_hsp2]{write_hsp2()}} writes a HSP data.frame to a file in TSV format.
}
