% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/nbinom.R
\name{cpp_rnbinom}
\alias{cpp_rnbinom}
\alias{plot_nbinom}
\title{Test negative_binomial_distribution in tekka simulation}
\usage{
cpp_rnbinom(n, size, mu)

plot_nbinom(n = 10000L)
}
\arguments{
\item{n}{number of observations.  If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{size}{target for number of successful trials, or dispersion
    parameter (the shape parameter of the gamma mixing distribution).
    Must be strictly positive, need not be integer.}

\item{mu}{alternative parametrization via mean: see \sQuote{Details}.}
}
\description{
Test negative_binomial_distribution in tekka simulation
}
\details{
\code{cpp_rnbinom} generates nbinom random numbers.

\code{plot_nbinom} draws distributions with various parameter sets.
}
\seealso{
stats::rnbinom
}
