% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{plot_parameters_json}
\alias{plot_parameters_json}
\title{Visualize parameter json file}
\usage{
plot_parameters_json(file = "")
}
\arguments{
\item{file}{path to a json file}
}
\description{
Visualize parameter json file
}
