% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{augment.sample_family}
\alias{augment.sample_family}
\alias{augment.genealogy}
\alias{layout_demography}
\alias{plot.genealogy}
\alias{plot.sample_family}
\title{Methods for quick visualization}
\usage{
\method{augment}{sample_family}(x, layout = NULL, ...)

\method{augment}{genealogy}(x, layout = NULL, ...)

layout_demography(x)

\method{plot}{genealogy}(x, ..., lwd = 0.5, cex = 5, pch = 16)

\method{plot}{sample_family}(x, ..., lwd = 0.5, cex = 5, pch = 16)
}
\arguments{
\item{x}{sample_family data.frame.}

\item{layout}{A data.frame.}

\item{...}{passed to layout function.}

\item{lwd}{passed to \link[ggplot2:geom_segment]{ggplot2::geom_segment}.}

\item{cex, pch}{passed to \link[ggplot2:geom_point]{ggplot2::geom_point} and \link[ggplot2:geom_text]{ggplot2::geom_text}.}
}
\description{
Methods for quick visualization
}
