% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{as_pop}
\alias{as_pop}
\alias{write_pop}
\alias{read_pop}
\title{POP format}
\usage{
as_pop(samples)

write_pop(x, path = "pop.txt")

read_pop(path)
}
\arguments{
\item{samples}{A \code{sample_family} data.frame of \code{\link[=tekka]{tekka()}} result.}

\item{x}{An outcome of \code{\link[=as_pop]{as_pop()}}}

\item{path}{A file name or connection to write to}
}
\value{
A data.frame with "pop" class and six columns:
\itemize{
\item \code{cohort}: birth year of offspring
\item \code{capture_year}: of parents
\item \code{capture_age}: of parents
\item \code{location}: where parents were sampled
\item \code{pops}: the number of parent-offspring pairs
\item \code{comps}: the number of possible comparisons
}
}
\description{
Parent-offspring pairs are counted between potential offspring cohort and
potential parents grouped by age, year, and location of sampling.
Adults are included in potential offspring.
}
\details{
\code{\link[=as_pop]{as_pop()}} converts a \code{sample_family} data frame to POP format.

\code{\link[=write_pop]{write_pop()}} writes a POP data.frame to a file.
}
