% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{as_pop}
\alias{as_pop}
\alias{write_pop}
\title{Convert a result data frame to POP format}
\usage{
as_pop(.tbl, min_adult_age = 5L)

write_pop(x, path = "pop.txt")
}
\arguments{
\item{.tbl}{A result data frame}

\item{min_adult_age}{integer}

\item{x}{An outcome of as_pop()}

\item{path}{A file name or connection to write to}
}
\value{
tibble with additional class name "pop"
}
\description{
Convert a result data frame to POP format

Write a POP data frame to a file
}
