% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop2.R
\name{as_pop2}
\alias{as_pop2}
\alias{write_pop2}
\alias{read_pop2}
\title{Extended POP format}
\usage{
as_pop2(samples)

write_pop2(x, path = "pop2.tsv")

read_pop2(path)
}
\arguments{
\item{samples}{A \code{sample_family} data.frame of \code{\link[=tekka]{tekka()}} result.}

\item{x}{An outcome of \code{\link[=as_pop2]{as_pop2()}}}

\item{path}{A file name or connection to write to.}
}
\value{
A data.frame with "pop" class and six columns:
\itemize{
\item \code{cohort_parent}, \code{cohort_offspring}: birth year of samples
\item \code{capture_age_parent}, \code{capture_age_offspring}
\item \code{location_parent}, \code{location_offspring}: of sampling
\item \code{pops}: the number of parent-offspring pairs
\item \code{comps}: the number of possible comparisons
}
}
\description{
Parent-offspring pairs are counted between potential offspring cohort and
potential parents grouped by birth year, capture age, location of sampling.
Adults are included in potential offspring.
}
\details{
\code{\link[=as_pop2]{as_pop2()}} converts a \code{sample_family} data frame to POP format.

\code{\link[=write_pop2]{write_pop2()}} writes a POP data.frame to a file.
}
