% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{make_snp}
\alias{make_snp}
\title{Make snp table from a result}
\usage{
make_snp(.tbl, lambda, rho = 0)
}
\arguments{
\item{.tbl}{result tibble}

\item{lambda}{mutation rate per haploid per generation}

\item{rho}{recombination rate per generation}
}
\description{
Make snp table from a result
}
