% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{make_snp}
\alias{make_snp}
\alias{make_snp_chromosome}
\alias{place_mutations}
\title{Functions to generate SNPs on given samples.}
\usage{
make_snp(samples, ss)

make_snp_chromosome(genealogy, segsites)

place_mutations(genealogy, segsites, v_sampled = NULL)
}
\arguments{
\item{samples}{A data.frame: \code{sample_family} or output from \code{\link[=gather_segments]{gather_segments()}}.}

\item{ss}{A sequence of \code{segsites};
its length is the number of chromosome;
each element is the number of segsites on a chromosome.
If a named vector is given, the output is also named.}

\item{genealogy}{An output from \code{\link[=make_gene_genealogy]{make_gene_genealogy()}}.}

\item{segsites}{The number of segregating sites on a segment.}

\item{v_sampled}{The sampled vertices. Use this to fix the output order.}
}
\description{
Functions to generate SNPs on given samples.
}
\details{
\code{\link[=make_snp]{make_snp()}} generates a list of SNP data.frames by calling
\code{\link[=make_snp_chromosome]{make_snp_chromosome()}} in parallel.
Use \code{RNGkind("L'Ecuyer-CMRG")} and \code{set.seed()} to get reproducible results.
The number of CPU cores used can be configured via \code{mc.cores} option.

\code{\link[=make_snp_chromosome]{make_snp_chromosome()}} simulate a SNP data.frame on a chromosome by calling
\code{\link[=place_mutations]{place_mutations()}} and \code{recombination()}.

\code{\link[=place_mutations]{place_mutations()}} generates a SNP data.frame by randomly placing a fixed
number of mutations on a given genealogy.
It means that all the sites are perfectly linked with each other.
}
