% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genealogy.R, R/snp.R
\name{annotate_sampled}
\alias{annotate_sampled}
\alias{make_snp}
\alias{make_snp_chromosome}
\alias{place_mutations}
\title{Functions to generate SNP matrix}
\usage{
annotate_sampled(genealogy)

make_snp(samples, ss = c(2L, 2L))

make_snp_chromosome(genealogy, segsites)

place_mutations(genealogy, segsites, v_sampled = NULL)
}
\arguments{
\item{genealogy}{An output from \code{\link[=make_gene_genealogy]{make_gene_genealogy()}}.}

\item{samples}{A \code{sample_family} data.frame.}

\item{ss}{A sequence of \code{segsites};
its length is the number of segments;
each element is the number of segsites on a segment.}

\item{segsites}{The number of segregating sites on a segment.}

\item{v_sampled}{The sampled vertices. Use this to fix the output order.}
}
\description{
Functions to generate SNP matrix
}
\details{
\code{\link[=annotate_sampled]{annotate_sampled()}} adds "sampled" edge attribute: \code{TRUE} if sampled,
\code{FALSE} if upstream of samples, \code{NA} if unrelated.

\code{\link[=make_snp]{make_snp()}} is a shortcut to generate a SNP matrix from
a \code{sample_family} data.frame.

\code{\link[=make_snp_chromosome]{make_snp_chromosome()}} simulate a SNP matrix on a chromosome by calling
\code{\link[=place_mutations]{place_mutations()}} and \code{recombination()}.

\code{\link[=place_mutations]{place_mutations()}} generates a SNP matrix by randomly placing a fixed
number of mutations on a given genealogy.
}
