% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{make_snp}
\alias{make_snp}
\alias{gather_segments}
\alias{make_gene_genealogy}
\alias{count_uncoalesced}
\title{Functions to generate gene genealogy and SNP matrix}
\usage{
make_snp(.tbl, ss = c(2L, 2L))

gather_segments(.tbl)

make_gene_genealogy(segments, segsites = 0L)

count_uncoalesced(genealogy)
}
\arguments{
\item{.tbl}{\code{sample_family} data.frame.}

\item{ss}{A sequence of \code{segsites};
its length is the number of segments;
each element is the number of segsites on a segment.}

\item{segments}{An output from \code{\link[=gather_segments]{gather_segments()}}.}

\item{segsites}{The number of segregating sites on a segment.}

\item{genealogy}{An output from \code{\link[=make_gene_genealogy]{make_gene_genealogy()}}.}
}
\description{
Functions to generate gene genealogy and SNP matrix
}
\details{
\code{\link[=make_snp]{make_snp()}} generates a SNP matrix.

\code{\link[=gather_segments]{gather_segments()}} transform individual-based \code{sample_family} into
segment-based table.

\code{\link[=make_gene_genealogy]{make_gene_genealogy()}} generates random gene genealogy.

\code{\link[=count_uncoalesced]{count_uncoalesced()}} counts uncoalesced lineages in a gene genealogy
}
