% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{make_snp}
\alias{make_snp}
\title{Functions to process SNPs}
\usage{
make_snp(.tbl, segsites = 1L)
}
\arguments{
\item{.tbl}{sample_family}

\item{segsites}{number of segregating sites}
}
\description{
Functions to process SNPs
}
\details{
\code{make_snp} makes snp table from a result.
}
