% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{make_snp}
\alias{make_snp}
\alias{gather_segments}
\alias{make_gene_genealogy}
\alias{place_mutations}
\alias{count_uncoalesced}
\alias{annotate_sampled}
\title{Functions to generate gene genealogy and SNP matrix}
\usage{
make_snp(.tbl, ss = c(2L, 2L))

gather_segments(.tbl)

make_gene_genealogy(segments)

place_mutations(genealogy, segsites, v_sampled = NULL)

count_uncoalesced(genealogy)

annotate_sampled(genealogy)
}
\arguments{
\item{.tbl}{\code{sample_family} data.frame.}

\item{ss}{A sequence of \code{segsites};
its length is the number of segments;
each element is the number of segsites on a segment.}

\item{segments}{An output from \code{\link[=gather_segments]{gather_segments()}}.}

\item{genealogy}{An output from \code{\link[=make_gene_genealogy]{make_gene_genealogy()}}.}

\item{segsites}{The number of segregating sites on a segment.}

\item{v_sampled}{The sampled vertices. Use this to fix the output order.}
}
\description{
Functions to generate gene genealogy and SNP matrix
}
\details{
\code{\link[=make_snp]{make_snp()}} is a shortcut to generate a SNP matrix from
a \code{sample_family} data.frame.

\code{\link[=gather_segments]{gather_segments()}} transforms an individual-based \code{sample_family} into
a segment-based table.

\code{\link[=make_gene_genealogy]{make_gene_genealogy()}} generates a random gene genealogy from a
segment-based table.

\code{\link[=place_mutations]{place_mutations()}} generates a SNP matrix by randomly placing a fixed
number of mutations on a given genealogy.

\code{\link[=count_uncoalesced]{count_uncoalesced()}} counts uncoalesced roots of the samples.

\code{\link[=annotate_sampled]{annotate_sampled()}} adds "sampled" edge attribute: \code{TRUE} if sampled,
\code{FALSE} if upstream of samples, \code{NA} if unrelated.
}
