% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tekka.R
\name{tekka}
\alias{tekka}
\title{Run C++ simulation.}
\usage{
tekka(args = character(0L), cache = FALSE)
}
\arguments{
\item{args}{command line arguments as a string vector.}

\item{cache}{use cache if TRUE.}
}
\description{
Run C++ simulation.
}
