% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/tekka.R
\name{cpp_tekka}
\alias{cpp_tekka}
\alias{tekka}
\title{Run C++ simulation.}
\usage{
cpp_tekka(args)

tekka(args = character(0L))
}
\arguments{
\item{args}{command line arguments as a string vector.}
}
\description{
Run C++ simulation.
}
\details{
\code{cpp_tekka} is a Rcpp function to call tekka
}
