% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tekka.R
\name{tekka}
\alias{tekka}
\title{Run the \code{tekka} simulator.}
\usage{
tekka(args = character(0L), cache = NULL)
}
\arguments{
\item{args}{Command line arguments as a string vector.
See \href{https://heavywatal.github.io/tekka/group__parameters.html}{\code{tekka} manual}
and \code{tekka("--help")} for available options.
Note that a random \code{--seed} is appended if not given.
Use \code{set.seed()} or \code{--seed} explicitly for reproducibility and caching.}

\item{cache}{Parent directory for tekka output.
\code{TRUE} is equivalent to "." (current directory).
\code{\link[=tempdir]{tempdir()}} is used in other cases including the default (\code{NULL}),
which is discarded at the end of an R session.
Set \code{FALSE} to force \code{tekka} to run and overwrite previous results if any.}
}
\value{
A data.frame read by \code{\link[=read_tekka]{read_tekka()}}.
}
\description{
A result is first written to a directory in \code{cache},
and then read into a data.frame.
If a previous result with exactly the same arguments is found in \code{cache},
it is read without calling \code{tekka}.
}
\examples{
set.seed(666)
# tekka("--help")

result = tekka("-y25 -l2 --sa 2,2 --sj 2,2")
result |> dplyr::select(!"outdir")

result$sample_family[[1L]]

result$demography[[1L]]
}
\seealso{
\code{\link[=tekka_path]{tekka_path()}} to get information about the \code{tekka} installation.
}
