% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tekkamaki-package.R
\docType{package}
\name{tekkamaki-package}
\alias{tekkamaki}
\alias{tekkamaki-package}
\title{tekkamaki: Individual-based simulator of pacific bluefin tuna}
\description{
Individual-based simulator of pacific bluefin tuna.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/heavywatal/tekkamaki}
  \item Report bugs at \url{https://github.com/heavywatal/tekkamaki/issues}
}

}
\author{
\strong{Maintainer}: Watal M. Iwasaki \email{heavy.watal@gmail.com} (0000-0002-4718-7605)

}
\keyword{internal}
