% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{as_vcf}
\alias{as_vcf}
\alias{write_vcf}
\alias{read_vcf}
\alias{as_vcf_gt}
\alias{add_pos_id}
\title{Utilities for VCF format}
\usage{
as_vcf(x, phased = TRUE, chrom = NA_character_, pos = NA_integer_)

write_vcf(x, file)

read_vcf(file)

as_vcf_gt(x, phased = TRUE)

add_pos_id(x)
}
\arguments{
\item{x}{SNP data.frame or VCF data.frame.}

\item{phased}{A logical to switch separators: \code{|} vs \code{/}.}

\item{chrom}{Characters.}

\item{pos}{Integers.}

\item{file}{A path or connection.}
}
\description{
\code{\link[=as_vcf]{as_vcf()}} converts a SNP data.frame to a VCF-like data.frame.
}
