% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{as_vcf}
\alias{as_vcf}
\alias{add_pos_id}
\alias{write_vcf}
\alias{read_vcf}
\alias{as_vcf_gt}
\title{Utilities for VCF format}
\usage{
as_vcf(x, phased = TRUE, chrom = NA_character_, pos = NA_integer_)

add_pos_id(x)

write_vcf(x, file)

read_vcf(file)

as_vcf_gt(x, phased = TRUE)
}
\arguments{
\item{x}{A SNP data.frame or VCF data.frame.}

\item{phased}{A logical to switch separators: \code{|} vs \code{/}.}

\item{chrom}{Characters.}

\item{pos}{Integers.}

\item{file}{A path or connection.}
}
\value{
A VCF-like data.frame with "vcf" subclass.

\code{\link[=as_vcf_gt]{as_vcf_gt()}} returns a single-column data.frame with "vcf_gt" subclass.
}
\description{
\code{\link[=as_vcf]{as_vcf()}} converts a SNP data.frame generated by \code{\link[=make_snp]{make_snp()}} to a VCF-like data.frame.
Only the "CHROM", "FORMAT", and "gt" columns are filled.

\code{\link[=add_pos_id]{add_pos_id()}} is useful to fill the "POS" and "ID" columns.
it assigns sequential integers to the "POS" column for each "CHROM",
and combines it with "CHROM" to make a unique "ID" values.

\code{\link[=write_vcf]{write_vcf()}} and \code{\link[=read_vcf]{read_vcf()}} are used to write and read VCF files.
If a SNP data.frame is given, it is converted before writing.
}
\examples{
set.seed(666)
result = tekka("-y20 -l2 --sa 2,2 --sj 2,2")
samples = result$sample_family[[1L]]
ss = c(chr1 = 3, chr2 = 2)
snp = make_snp(samples, ss)
vcf = as_vcf(snp) |> add_pos_id()
vcf
}
\seealso{
\code{\link[=make_snp]{make_snp()}} to generate SNP data.frames.

\url{https://samtools.github.io/hts-specs/} for VCF format specification.
}
